create retrofit data classes

+++ app/src/main/java/name/mharbovskyi/findchargingstation/domain/usecase/oauth/Models.kt
@@ -3,7 +3,7 @@ package name.mharbovskyi.findchargingstation.domain.usecase.oauth
 data class AuthTokens(
     val accessToken: String,
     val refreshToken: String,
-    val expirationTimeSec: Int
+    val expirationTimeSec: Long
 )
 
 val EMPTY_TOKENS = AuthTokens("", "", 0)
